# schemas/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CapacityUnit": {
        "enum": [
          "people",
          "beds",
          "teams",
          "people_day"
        ],
        "title": "CapacityUnit",
        "type": "string"
      },
...
      "DataStatus": {
        "enum": [
          "observed",
          "reanalysis",
          "reported",
          "modeled",
          "mixed",
          "synthetic_fixture",
          "synthetic_demo",
          "algorithmic_result",
          "scenario_model"
        ],
        "title": "DataStatus",
        "type": "string"
      },
      "EvaluationRequest": {
        "additionalProperties": false,
        "properties": {
          "at_risk_population": {
            "default": 100000,
            "exclusiveMinimum": 0,
            "title": "At Risk Population",
            "type": "integer"
          },
          "hazard_threshold": {
            "default": 0.5,
            "maximum": 1,
            "minimum": 0,
            "title": "Hazard Threshold",
            "type": "number"
          }
        },
        "title": "EvaluationRequest",
        "type": "object"
      },
      "EvaluationResponse": {
        "additionalProperties": false,
        "properties": {
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "at_risk_population": {
            "exclusiveMinimum": 0,
            "title": "At Risk Population",
            "type": "integer"
          },
          "by_type": {
            "items": {
              "$ref": "#/components/schemas/FacilityTypeEvaluation"
            },
            "title": "By Type",
            "type": "array"
          },
          "covered_population_per_budget_point": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Covered Population Per Budget Point"
          },
          "data_status": {
            "const": "scenario_model",
            "default": "scenario_model",
            "title": "Data Status",
            "type": "string"
          },
          "facility_count": {
            "minimum": 0,
            "title": "Facility Count",
            "type": "integer"
          },
          "modeled_coverage_ratio": {
            "maximum": 1,
            "minimum": 0,
            "title": "Modeled Coverage Ratio",
            "type": "number"
          },
          "modeled_covered_population": {
            "minimum": 0,
            "title": "Modeled Covered Population",
            "type": "integer"
          },
          "modeled_uncovered_population": {
            "minimum": 0,
            "title": "Modeled Uncovered Population",
            "type": "integer"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "total_budget_points": {
            "minimum": 0,
            "title": "Total Budget Points",
            "type": "integer"
          }
        },
        "required": [
          "scenario_id",
          "facility_count",
          "at_risk_population",
          "modeled_covered_population",
          "modeled_uncovered_population",
          "modeled_coverage_ratio",
          "total_budget_points",
          "assumptions"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
...
      "FacilityCreate": {
        "additionalProperties": false,
        "properties": {
          "budget_points": {
            "anyOf": [
              {
                "maximum": 5,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Budget Points"
          },
          "capacity_unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CapacityUnit"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "capacity_value": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Capacity Value"
          },
          "lat": {
            "maximum": 90,
            "minimum": -90,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180,
            "minimum": -180,
            "title": "Lon",
            "type": "number"
          },
          "service_radius_km": {
            "anyOf": [
              {
                "exclusiveMinimum": 0,
                "maximum": 200,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Service Radius Km"
          },
          "type": {
            "$ref": "#/components/schemas/FacilityType"
          }
        },
        "required": [
          "type",
          "lon",
          "lat"
        ],
        "title": "FacilityCreate",
        "type": "object"
      },
...
      "FacilityRead": {
        "additionalProperties": false,
        "properties": {
          "budget_points": {
            "anyOf": [
              {
                "maximum": 5,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Budget Points"
          },
          "capacity_unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CapacityUnit"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "capacity_value": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Capacity Value"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_simulated": {
            "title": "Is Simulated",
            "type": "boolean"
          },
          "lat": {
            "maximum": 90,
            "minimum": -90,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180,
            "minimum": -180,
            "title": "Lon",
            "type": "number"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "service_radius_km": {
            "anyOf": [
              {
                "exclusiveMinimum": 0,
                "maximum": 200,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Service Radius Km"
          },
          "type": {
            "$ref": "#/components/schemas/FacilityType"
          }
        },
        "required": [
          "type",
          "lon",
          "lat",
          "id",
          "scenario_id",
          "is_simulated"
        ],
        "title": "FacilityRead",
        "type": "object"
      },
      "FacilityType": {
        "enum": [
          "shelter",
          "medical",
          "rescue",
          "warehouse"
        ],
        "title": "FacilityType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScenarioCreate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ScenarioCreate",
        "type": "object"
      },
      "ScenarioRead": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "ScenarioRead",
        "type": "object"
      },
      "StormCatalogResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0,
            "title": "Count",
            "type": "integer"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StormSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          }
        },
        "required": [
          "data_status",
          "items",
          "count"
        ],
        "title": "StormCatalogResponse",
        "type": "object"
      },
      "StormDetail": {
        "additionalProperties": false,
        "properties": {
          "ace": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Ace"
          },
          "basin": {
            "$ref": "#/components/schemas/Basin"
          },
          "classic": {
            "default": false,
            "title": "Classic",
            "type": "boolean"
          },
          "classic_rank": {
            "anyOf": [
              {
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Classic Rank"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "duration_hours": {
            "minimum": 0,
            "title": "Duration Hours",
            "type": "number"
          },
          "end_time": {
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "id": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "impact": {
            "$ref": "#/components/schemas/StormImpact"
          },
          "impact_available": {
            "default": false,
            "title": "Impact Available",
            "type": "boolean"
          },
          "impact_score": {
            "anyOf": [
              {
                "maximum": 100,
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Impact Score"
          },
          "landfall_count": {
            "default": 0,
            "minimum": 0,
            "title": "Landfall Count",
            "type": "integer"
          },
          "max_wind_ms": {
            "anyOf": [
              {
                "maximum": 150,
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Max Wind Ms"
          },
          "min_pressure_hpa": {
            "anyOf": [
              {
                "maximum": 1100,
                "minimum": 800,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Min Pressure Hpa"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "reported_damage_usd_2024": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Reported Damage Usd 2024"
          },
          "reported_deaths": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Reported Deaths"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "score_coverage": {
            "default": 0,
            "maximum": 1,
            "minimum": 0,
            "title": "Score Coverage",
            "type": "number"
          },
          "season": {
            "maximum": 2200,
            "minimum": 1840,
            "title": "Season",
            "type": "integer"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "track": {
            "items": {
              "$ref": "#/components/schemas/TrackPoint"
            },
            "minItems": 1,
            "title": "Track",
            "type": "array"
          },
          "wind_available": {
            "default": false,
            "title": "Wind Available",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "season",
          "basin",
          "start_time",
          "end_time",
          "duration_hours",
          "data_status",
          "track",
          "impact"
        ],
        "title": "StormDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "高影响热带气旋可视分析与台湾设施情景系统的本地 API 骨架。",
    "title": "CycloneScope API",
    "version": "0.1.0-local"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scenarios": {
      "get": {
        "operationId": "list_scenarios_api_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioRead"
                  },
                  "title": "Response List Scenarios Api Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "scenarios"
        ]
      },
      "post": {
        "operationId": "create_scenario_api_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/{scenario_id}/evaluate": {
      "post": {
        "operationId": "evaluate_scenario_api_scenarios__scenario_id__evaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/{scenario_id}/facilities": {
      "post": {
        "operationId": "add_facility_api_scenarios__scenario_id__facilities_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacilityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilityRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Facility",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/storms": {
      "get": {
        "operationId": "storms_api_storms_get",
        "parameters": [
          {
            "in": "query",
            "name": "basin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Basin"
            }
          },
          {
            "in": "query",
            "name": "classic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classic"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "impact_score",
              "pattern": "^(impact_score|season|max_wind_ms)$",
              "title": "Sort By",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StormCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Storms",
        "tags": [
          "storms"
        ]
      }
    },
    "/api/storms/{storm_id}": {
      "get": {
        "operationId": "storm_detail_api_storms__storm_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "storm_id",
            "required": true,
            "schema": {
              "title": "Storm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StormDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Storm Detail",
        "tags": [
          "storms"
        ]
      }
    },
    "/api/storms/{storm_id}/impact/summary": {
      "get": {
        "operationId": "storm_impact_api_storms__storm_id__impact_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "storm_id",
            "required": true,
            "schema": {
              "title": "Storm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StormImpactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Storm Impact",
        "tags": [
          "impact"
        ]
      }
    },
    "/api/storms/{storm_id}/track": {
      "get": {
        "operationId": "storm_track_api_storms__storm_id__track_get",
        "parameters": [
          {
            "in": "path",
            "name": "storm_id",
            "required": true,
            "schema": {
              "title": "Storm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StormTrackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Storm Track",
        "tags": [
          "storms"
        ]
      }
    },
    "/api/storms/{storm_id}/wind/manifest": {
      "get": {
        "operationId": "wind_manifest_api_storms__storm_id__wind_manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "storm_id",
            "required": true,
            "schema": {
              "title": "Storm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wind Manifest",
        "tags": [
          "wind"
        ]
      }
    },
    "/api/storms/{storm_id}/wind/sample-frame": {
      "get": {
        "operationId": "sample_wind_frame_api_storms__storm_id__wind_sample_frame_get",
        "parameters": [
          {
            "in": "path",
            "name": "storm_id",
            "required": true,
            "schema": {
              "title": "Storm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindFrame"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sample Wind Frame",
        "tags": [
          "wind"
        ]
      }
    }
  }
}